+++ server/src/data/config.json
@@ -6,7 +6,7 @@
     "min": 5,
     "max": 50,
     "step": 0.2,
-    "value": 15.2
+    "value": 17
   },
   {
     "type": "size",
@@ -15,7 +15,7 @@
     "min": 5,
     "max": 25,
     "step": 0.2,
-    "value": 6.8
+    "value": 14
   },
   {
     "type": "frame",
