+++ server/src/data/select.json
@@ -1,6 +1,6 @@
 {
-  "listValue": "Лист-7 0.3 ширина 1м",
-  "pipeValue": "Труба 20х20",
-  "choiceOfFrame": "standard",
-  "choiceOfMaterial": "metal"
+  "listValue": "Лист-1 0.5 ширина 1.8м",
+  "pipeValue": "Труба 30х30",
+  "choiceOfFrame": "strong",
+  "choiceOfMaterial": "plastic"
 }
