+++ src/components/InputForm/InputForm.tsx
@@ -1,92 +1,31 @@
 import style from './InputForm.module.css';
 import {
-    ConfigurationApiType,
-    MaterialsApiType,
-    SelectApiType,
     useGetConfigQuery,
     useGetMaterialsQuery,
     useGetSelectOptionsQuery,
     useUpdateConfigMutation,
-    useUpdateSelectOptionsMutation
 } from "../../redux/services/calculatorApi.ts";
 import {AllOrNone} from "../../type/Types.ts";
 import {AComponentUsingInput} from "../../lib/Input/Input.tsx";
 import {useEffect, useState} from "react";
+import {ConfigurationApiType, MaterialsApiType} from "../../type/zodTypes.ts";
+import {useSelectChange} from "../../lib/Hooks/useSelectChange.ts";
+import {useOptionSelect} from "../../lib/Hooks/useOptionSelect.tsx";
+import {useGetSizeConfig} from "../../lib/Hooks/useGetSizeConfig.ts";
 
 
 function InputForm() {
 
-    const {data: materials, error: materialsError, isLoading: materialsIsLoading} = useGetMaterialsQuery();
+    const { error: materialsError, isLoading: materialsIsLoading} = useGetMaterialsQuery();
     const {data: config, error: configError, isLoading: configIsLoading} = useGetConfigQuery();
     const {data: selectValue} = useGetSelectOptionsQuery()
 
+    const [updateConfig] = useUpdateConfigMutation();
+    const {optionSelect} = useOptionSelect();
+    const {getSizeConfig} = useGetSizeConfig();
+
     const [choiceOfMaterial, setChoiceOfMaterial] = useState('');
     const [choiceOfFrame, setChoiceOfFrame] = useState('');
-
-    const [updateConfig] = useUpdateConfigMutation();
-    const [updateSelectOptions] = useUpdateSelectOptionsMutation();
-
-
-    const handleSelectChange = (event: React.ChangeEvent<HTMLSelectElement>, type: OptionSelectType) => {
-        const selectValue = event.target.value;
-        switch (type) {
-            case "list":
-                updateSelectOptions({listValue: selectValue});
-                break;
-            case "pipe":
-                updateSelectOptions({pipeValue: selectValue});
-                break;
-            case "material": {
-                const selectedKey = config?.find(({name}) => name === selectValue);
-                if (selectedKey?.key) {
-                    setChoiceOfMaterial(selectValue)
-                    updateSelectOptions({
-                        choiceOfMaterial: selectedKey.key as SelectApiType['choiceOfMaterial']
-                    });
-                }
-                break;
-            }
-            case "frame": {
-                const selectedKey = config?.find(({name}) => name === selectValue);
-                if (selectedKey?.key) {
-                    setChoiceOfFrame(selectValue)
-                    updateSelectOptions({
-                        choiceOfFrame: selectedKey.key as SelectApiType['choiceOfFrame']
-                    });
-                }
-                break;
-            }
-        }
-    };
-
-    const optionSelect = (type: OptionSelectType) => {
-        switch (type) {
-            case 'list':
-            case 'pipe':
-            case 'fix':
-                return materials?.filter((matetial) => matetial.type === type).map((item, index) => {
-                    return <option key={index} value={item.name}>{item.name}</option>
-                })
-            case "frame":
-            case "material":
-            case "size":
-                return config?.filter((el) => el.type === type).map((item, index) => {
-                    return <option key={index} value={item.name}>{item.name}</option>
-                })
-
-        }
-    }
-
-    const getSizeConfig = (key: SizeConfigType): GetSizeConfigType | undefined => {
-        const sizeConfig = config?.find((item) => item.type === 'size' && item.key === key);
-        return sizeConfig ? {
-            min: sizeConfig.min ?? 0,
-            max: sizeConfig.max ?? 0,
-            step: sizeConfig.step ?? 0,
-            value: sizeConfig.value ?? 0,
-        } as GetSizeConfigType : undefined;
-    }
-
     const [form, setForm] = useState({
         width: getSizeConfig("width")?.min ?? 0,
         length: getSizeConfig("length")?.value ?? 0,
@@ -94,6 +33,8 @@ function InputForm() {
         pipe: "",
         frame: "",
     });
+
+    const {handleSelectChange} = useSelectChange(setChoiceOfFrame, setChoiceOfMaterial);
 
     useEffect(() => {
         setForm((prev) => ({
@@ -187,8 +128,8 @@ function InputForm() {
 export default InputForm;
 
 // types
-type OptionSelectType = MaterialsApiType['type'] | ConfigurationApiType['type']
-type SizeConfigType = ConfigurationApiType['key']
+export type OptionSelectType = MaterialsApiType['type'] | ConfigurationApiType['type']
+export type SizeConfigType = ConfigurationApiType['key']
 
 export type GetSizeConfigType = AllOrNone<ConfigurationApiType, 'min' | 'max' | 'step' | 'value'>
